Se hace un refactor de los paquetes del plugin
