+++ pom.xml
@@ -4,8 +4,12 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.cinepolis</groupId>
+    <groupId>mx.qbits.keycloak.plugin</groupId>
+    
+    <!-- OJO !!! CAMBIAR: auth-provider-cinepolis POR: auth-provider-xyz SEGÚN SE NECESITE-->
     <artifactId>auth-provider-cinepolis</artifactId>
+    <!-- FIN DE COMENTARIO -->
+    
     <version>1.0-SNAPSHOT</version>
 
     <properties>
@@ -62,7 +66,7 @@
                     <!-- MainClass in mainfest make a executable jar -->
                     <archive>
                       <manifest>
-                        <mainClass>cinepolis.keycoak.user.ManageProperties</mainClass>
+                        <mainClass>mx.qbits.keycloak.plugin.ManageProperties</mainClass>
                       </manifest>
                     </archive>
