+++ src/main/resources/META-INF/services/org.keycloak.storage.UserStorageProviderFactory
@@ -1,1 +1,1 @@
-cinepolis.keycoak.user.CustomUserStorageProviderFactory
+mx.qbits.keycloak.plugin.CustomUserStorageProviderFactory
