+++ src/main/resources/database.properties
@@ -5,7 +5,10 @@ db.username=db-admin
 db.password=UrbiEtOrbi1
 
 # OJO *********************
-# NO OLVIDAR CAMBIAR LA LINEA 8 EN EL POM.XML QUE LLEVA EL NOMBRE DEL JAR
+# NO OLVIDAR CAMBIAR:
+#    1) LA LINEA 10 EN EL POM.XML QUE LLEVA EL NOMBRE DEL JAR
+#    2) En la linea 3 de este archivo reapuntar la base de datos
+#    3) En el docker del servidor reapuntar las variables de ambiente adecuadamente
 # OJO *********************
 
 driver.class.name=org.mariadb.jdbc.Driver
